Brain_Tumor/api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.25.2
Pillow==10.0.0
tensorflow==2.14.0
werkzeug==2.3.7
scikit-learn==1.3.0
opencv-python==4.8.0
matplotlib==3.7.2
grad-cam==1.4.8